stable version

+++ KCF/fhog.py
@@ -190,7 +190,7 @@ def getFeatureMaps(image, k, mapp):
                     width, numChannels)  # with @jit
     # ~0.001s
 
-    nearest = np.ones((k), np.int)
+    nearest = np.ones((k), np.int32)
     nearest[0:k // 2] = -1
 
     w = np.zeros((k, 2), np.float32)
